+++ components/timeline/Timeline.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import classNames from 'classnames';
-
+import LoadingOutlined from '@ant-design/icons/LoadingOutlined';
 import { ConfigContext } from '../config-provider';
 import type { TimelineItemProps } from './TimelineItem';
 import TimelineItem from './TimelineItem';
@@ -38,6 +38,8 @@ const Timeline: CompoundedComponent = (props) => {
     rootClassName,
     reverse = false,
     mode = '' as TimelineProps['mode'],
+    pendingDot,
+    items,
     ...restProps
   } = props;
   const prefixCls = getPrefixCls('timeline', customizePrefixCls);
@@ -50,7 +52,52 @@ const Timeline: CompoundedComponent = (props) => {
   // Style
   const [wrapSSR, hashId] = useStyle(prefixCls);
 
-  const { items: mergedItems, hasLabelItem } = useItems(props, prefixCls);
+  const mergedItems: TimelineItemProps[] = useItems(items, children);
+  const getPositionCls = (position: string, idx: number) => {
+    if (mode === 'alternate') {
+      if (position === 'right') return `${prefixCls}-item-right`;
+      if (position === 'left') return `${prefixCls}-item-left`;
+      return idx % 2 === 0 ? `${prefixCls}-item-left` : `${prefixCls}-item-right`;
+    }
+    if (mode === 'left') return `${prefixCls}-item-left`;
+    if (mode === 'right') return `${prefixCls}-item-right`;
+    if (position === 'right') return `${prefixCls}-item-right`;
+    return '';
+  };
+  const pendingNode = typeof pending === 'boolean' ? null : pending;
+
+  if (pending) {
+    mergedItems.push({
+      pending: !!pending,
+      dot: pendingDot || <LoadingOutlined />,
+      content: pendingNode,
+    });
+  }
+
+  if (reverse) {
+    mergedItems.reverse();
+  }
+  const itemsCount = mergedItems.length;
+  const lastCls = `${prefixCls}-item-last`;
+
+  const itemsList = mergedItems
+    .filter((item: TimelineItemProps) => !!item)
+    .map((item: TimelineItemProps, idx: number) => {
+      const pendingClass = idx === itemsCount - 2 ? lastCls : '';
+      const readyClass = idx === itemsCount - 1 ? lastCls : '';
+      return (
+        <TimelineItem
+          className={classNames([
+            className,
+            !reverse && !!pending ? pendingClass : readyClass,
+            getPositionCls(item?.position ?? '', idx),
+          ])}
+          {...item}
+          key={item.content as string}
+        />
+      );
+    });
+  const hasLabelItem = mergedItems.some((item: TimelineItemProps) => !!item?.label);
 
   const classString = classNames(
     prefixCls,
@@ -66,11 +113,9 @@ const Timeline: CompoundedComponent = (props) => {
     hashId,
   );
 
-  delete restProps.items;
-
   return wrapSSR(
     <ul {...restProps} className={classString}>
-      {mergedItems}
+      {itemsList}
     </ul>,
   );
 };
